GNN/requirements.txt
torch-geometric
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv

pandas

scikit-learn
